+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.5.3
+
+- added --allow-changes option to cobaya-run to allow changes in the input file
+- Updates for deprecation warnings
+- Minor optimization refactor and doc update
+
 ## 3.5.2
 
 - Updates for numpy 2 and other compatibility fixes